full preciion

+++ examples/symmetric_conv2d_vae/updated/lightning.py
@@ -80,7 +80,8 @@ class CVAE(pl.LightningModule):
         x = outputs['expected']
         x = x
         kld_loss = self.model.kld_loss().float()
-        recon_loss = self.criterion(recon_x.half(), x.half()).float()
+        # recon_loss = self.criterion(recon_x.half(), x.half()).float()
+        recon_loss = self.criterion(recon_x, x)
         loss = 1.0 * recon_loss + kld_loss
         # loss = recon_loss
         # only use when  on dp
@@ -98,9 +99,9 @@ class CVAE(pl.LightningModule):
         _, recon_x = outputs['forward']
         # recon_x = recon_x.clamp(0, 1)
         x = outputs['expected']
-        x = x.half()
-        kld_loss = self.model.kld_loss().float()
-        recon_loss = self.criterion(recon_x, x).float()
+        # x = x.half()
+        kld_loss = self.model.kld_loss()
+        recon_loss = self.criterion(recon_x, x)
         loss = 1.0 * recon_loss + kld_loss
         # loss = recon_loss
         # only use when  on dp
